Bug 694882: sanitize component mappings
PCLR, CMAP and CDEF data aren't properly sanitized - even after the
fixes from the last fuzzing round. This patch unifies all the required
checks in a new function opj_jp2_check_color which is invoked before
the cdef and cmap values are used.

This can be seen e.g. in:

66ea31acbb0f23a2bbc91f64d69a03f5_signal_sigsegv_13937c0_7030_5725.pdf

Thanks to Mateusz Jurczyk and Gynvael Coldwind of the Google Security
Team for providing the example files.