Bug 694871: move libjpeg away from gc memory
Change the libjpeg interface code to use a chunk allocator instead of garbage
collected, "unstable" memory.

Previously, the libjpeg integration used "Postscript" memory, that is, memory
that is subject to garbage collection and save/restore operations.

This means that Postscript like:
/In (file) (r) file /DCTDecode filter def
save
<<
  ...
  /ImageSource In
  ...
>> image
restore
In closefile

would almost certainly cause a crash, since the memory allocated by libjpeg
during the decoding for the "image" operator would be freed by the "restore"
operator, *before* the "closefile" destroyed the filter stream and the
libjpeg context.

If we opt to always wrap the "default" heap allocator in a chunk allocator
then we should probably remove the chunk allocator instances created
specifically for libjpeg.

No cluster differences.