PDF interpreter - Don't set UseCIEColor unless actually required.
The PDF specification says that if a PDF file contains a page which uses
a ColorSpace Resource of DefautlGray, DefaultRGB or DefaultCMYK then this
should be regarded as a "request that such colors be systematically
transformed (remapped) into device-independent CIE-based color spaces.".
Essentially the PDF equivalent of UseCIEColor.

This does rather beg the question of why not specify the colours in a
device-independent space in the first place.....

In any event, in order to achieve this the PDF interpreter pulls some
tricks. We start by defining the Default* resources as the equivalent
Device* spaces and *always* define UseCIEColor to true. Thus whenever
we set a device color space the colour machinery tries to use instead the
appropriate Device* space. Because the color space machinery has an
optimisation to jump straight out when the requested space is the same as
the current space, this means that the UseCIEColor transformation should
have no effect.

However, pdfwrite now emits a warning if UseCIEColor is true, so we want
to avoid unconditionally setting UseCIEColor (I also find it horrifying that
we do this terrible hackery).

This commit borrows code used to set the Default* spaces if they exist on
a page, and uses it to detect such spaces before we call setpagedevice
for each page. We now only set UseCIEColor if the page actually uses one
of the Default* spaces and therefore requires it.

I did consider trying to substitute the defined colour space with the
Default* space whenever a colour space definition takes place and a Default*
space is defined, but gave up as there were simply too many places where
color spaces could be defined.

This code does (somewhat unexpectedly) produce some differences.
These are either progressions or very tiny differences, and I feel that
*not* setting UseCIEColor should be preferable anyway. (The real puzzle
is why some PostScript files seem to differ, but they only differ when
sent to pdfwrite and the PDF is then interpreed. Looks like UseCIEColor
has some strange behaviour with ICCBased spaces).

with pdfwrite:
altona_technical_1v2_x3.pdf - very minor colour shift
altona_technical_v20_x4 - progression on page 7
09-31.ps - difference in gray ramp in CIE space
09-34.ps - indetectable differences in some CIE colours

all devices:
Bug692783.pdf - indetectable shits in background
Catx5720.pdf - slight colour shifts in 3 graphics
1021_-_transparency-issue.pdf - tiny colour shifts
1141_-_background_shading_cropped.pdf - tiny colour shifts