pdfwrite - fix xref when eliding duplicate patterns
commit fe99ca5fd2e3b191a76c7f7d791b72f3603e9fea detected duplicate shading
patterns, and did not emit them into the PDF file, instead using the first
definition. Unfortunately, due to a mis-understanding of the existing code
the xref entry remained assigned, which led to invalid xref tables
being written.

This commit fixes the xref, and adds some (hopefully useful) comments to
the relevant functions to try and avoid a repetition.

No differences expected