PDF interpreter - attempt to recover from degenerate text matrix or CTM
Some PDF files set a degenerate matrix or text matrix, eg

0 0 0 0 0 0 Tm

This causes us a problem with currentpoint and .getpath, both of which
need to invert the CTM. If the CTM is all 0 then it (clearly) can't be
inverted.

The code here emits a warning, in the case of settextposition it simply
does not advance the current point (which seems correct, since the text
will make no changes). In the case where we attempt to grestore (Q) but
preserving the current path, we drop the path and make a new path. This
may well not be correct, but its hard to see what would be.

No differences expected.