pdfwrite - enable the EMBED pdfamrk
Bug #690043 "Can't create a pdf file with attachment"

This turned out to be more complex than I expected. I had planned to lever
off the /Dests code, but....

It seems that our Named destinations support is only PDF 1.1. The specification
was changed ion PDF 1.2 to use a name tree instead of a simple dictionary
and we didn't alter our code. The EmbeddedFiles wasn't supported in PDF 1.1
so it has always needed a name tree.

So in addition to adding the EMBED pdfmark, we now also will write a name
tree for Dests if the CompatibilityLevel is > 1.1. Note that the 'tree'
we write is the simplest possible one and is likely to be sub-optimal, we
write all the entries in the root of the tree. However its legal and should
work, we can extend this later if needs be.

I don't have a good way to test this, but the output looks OK by inspection
and we don't get any regressions.