pdfwrite - more EMBED pdfmrk changes
commit 3e5ae4ea39655643ae352cf4723702a164c10417 omitted a crucial change
which swapped the search order of /FS and /Name. Without that the dict
entries are not correctly named.

This oversight demonstrated a flaw in the error handling resulting in invalid
PDF files being written, which we correct here. If we get an error we close
the 'aside' so that we don't end up writing the PDF trailer to a temporary
file instead of the real PDF file.

No differences expected