PDF interpreter - disable Comment parsing for PDF files
Bug #694987 "PDF metadata messed up for PDF->PDF conversion"

This is caused by parsing DSC comments from an embedded font file. Of course
the font should not contain DSC comments, but it seems FontForge always
embeds such.

There is nothing useful to be gained from *PostScript* comments in a PDF
file, so the simplest solution is to disable comment parsing for the
duration of the PDF file.

This will also avoid problems with PostScript XObjects.

No differences expected