Fix Windows display device -dDisplayFormat=16#a0800 regression.
Commit 82fc3bd change to msvc.mak removed setting the GX_COLOR_INDEX_TYPE,
and code in base/gxcindex.h set this incorrectly to 'ulong'. Add the code
that was removed and also add conditional compile to set this from
ARCH_SIZEOF_GX_COLOR_INDEX if it is not defined.

Also add a 'compile time assert' that causes the build to fail if the
sizeof(GX_COLOR_INDEX_TYPE) is not equal to ARCH_SIZEOF_GX_COLOR_INDEX.
Thanks to Robin for finding this trick on the web.