Fix bug 694821. Ignore SeparationOrder with the psdcmykog device.
While this is a DeviceN device with some non-standard components
that can be painted using Separation or DeviceN colorspaces, this
device always outputs all 6 channels and can't deal with a subset
of the channels, so return an error (undefined) for SeparationOrder.
the separation_map remains constant to avoid color distortion if
SeparationOrder is used.

Also make sure that we never add separations (NO_AUTO_SPOT_COLORS)
from the get_color_comp_index procedure.