PS interpreter - Fix rotation problem with EPSFitPage
Bug #695016 "Gradient eps file to pdf"

When fitting an EPS Bounding Box to a specified page size, the code looks
at the width and height of both the page and the EPS file, to see if the
EPS will fit 'better' by rotating it.

In this particular case the page is square, but there was no special case
for this, which mean that landscape pages wold be rotated twice, causing
the content to disappear.

This commit special cases square media sizes and does not attempt to rotate
the EPS, no matter what its orientation, as there is clearly no point in
doing so. This prevents the double rotation and resolves the problem.

Tested with landscape, portrait and square media and EPS files, in all
combinations.

No differences expected