Bug 695027: use a consistent allocator in the FT server
There was some mixing of memory allocators in the Freetype integration code,
most of the code used the chunk allocator instance created during the
server initialization, but glyph data allocations were using the allocator from
the font object.

With two different FAPI servers in place, and with the right combination of
events, that could mean we'd lose the allocator pointer when destroying
the font object.

We're consistent in that all the memory allocations in the FAPI/Freetype code
are directed through the chunk allocator created for that purpose.

No cluster differences.