Bug #695034 Improve handling of word spacing with CIDFonta
Bug #695034 "Numbers and words are overlapping"

Word spacing does not apply to ti CIDFOnts, *except* for 1-byte mappings
where the byte 0x20 has word spacing applied to it.

The previous code attempted to check this by using a heuristic to determine
a single byte mapping in order to avoid adding more entries to the text
enumerator. unfortunately this file defeats that heuristic which meant
that it was necessary to calculate the number of bytes decoded when we
extract one character, and store that information so that the 'move'
routine can apply word spacing if required.

This shows two differences with a cluster run. Bug694436.pdf is very subtly
different, I'm assuming this is a progression. The Sumatra file
850_-_wrong_default_for_asian_fonts.pdf shows very significant improvements
with this commit.