ps2write - more colour conversion work
Bug #694774

When outputting to PostScript we don't preserve certain colour spaces which
either cannot be represented in PDF (CIEBased) or in level 2 PostScript
(ICCBased/Lab/DeviceN) are not preserved. These should be written as device
colours instead, but this wasn't always happening correctly.

This commit should resolve this, and should now make it possible (but still
inadvisable) to use -dUseCIEColor with both pdfwrite and ps2write.

A small number of files show (frankly imperceptible) colour shifts with
ps2write with this change.