Fix pngalpha when PDF has transparency. Bugs 687630, 693024 and 695042.
We needed a pngalpha_put_image procedure to properly collect the pdf14
compositor alpha value. Note that this is NOT simple 0 or 1 alpha, but
the actual effective page level opacity from the PDF. Note that some
viewers may not correctly display these despite this being part of
the PNG spec.