Fix Bug 694685 - Seg faults found by fuzzing in sfopen.
PCL would continue processing with a corrupt font.  This could lead to
bad memory accesses as demonstrated by the fuzzing example.  For now,
we return an error when a corrupt font is encountered and end the job.
Likely, a lighter approach is called for: continue process without
defining the font.  This would be a bit more involved and we'll
consider it if users report HP precedent for the behavior.