Bug 695007: Fix problems with unicode chars in TEMP path.
If there are unicode chars in the temporary path (either that
given by TMPDIR, TEMP or windows own GetTempPath functions)
when opening a scratch file, we mishandle them.

This commit improves the behaviour to treat paths consistently
as UTF8.

This incorporates changes from Paul Gardiner to make the winrtsup.cpp
functions use wide chars. This simplifies the functions and integrates
better with the calling code. Also fix some uses of sizeof where size
in chars was required.