pdfwrite - do not apply the 'Default*' colour spaces
Noticed while testing by Chris.

The PDF spec says that if any of DefaultGray, DefaultRGB or DefaultCMYK are
present in a PDF page, then we should use those to do a conversion to
device-independent colour space for colours specified in the matching Device*
space.

This is broadly equivalent to the UseCIEColor PostScript user parameter but
unlike that parameter the user has no control over it. Previously we applied
this conversion but it seems unreasonable to do this for pdfwrite as this
will convert a PDF using device spaces into a PDF using ICCBased spaces.

Accordingly we now test the device properties and for HigLevel devices we
do *not* perform this conversion.

If the conversion really is desired then setting -dUseCIEColor will perform it.

This results in a very few files showing differences, neither progressions
nor regressions, merely different. However the pdfwrite warning about use
of UseCIEColor does go away.