pdfwrite revert 3894e9c8a3ab125c82048da3bf81515500bb2da0
Bug #695083 " Regression: pdfwrite DEVICE generates extra page with -dLastPage= option"
This wasn't a customer bug report, and it doesn't work well with PCL, when
specifying -dLastPage, so I've decided just to revert it.

A very few PCL files are flagged by the cluster as different, but there are
no bmpcmp diffs.