pxlmono/pxlcolor: Transform deep (24-bit) images with an ICC transform to emit high-level images. Bug 690867.
Deep images with an associated rendering intent
or custom colour spaces, or icc profiles, could be emitted
more efficiently as high-level PXL images if the individual
pixels values are transformed through the icclink transform.

The new code path is off by default, and switched on via an option
(-diccTransform). When the new code is thoroughly tested, this
option will be removed and the new code path will become the default.

Cluster tesed - new code is optional and off by default,
no difference is expected.

Signed-off-by: Hin-Tak Leung <hintak@ghostscript.com>