PDF interpreter - fix multiple xref stream reading
Bug #695086 "**** Error: Trailer is not found."

When reading Xref streams, the PDF interpreter resizes its internal xref
tracking objects by looking at the /Size of the stream and making sure
the objects are at least that big. However it did not account for the fact
that an Xref stream may not start with object 0.

In this case the first stream encountered had a /Size of 2, and an /Index
of [6 0]. This meant that we allocated objects of size 2, then tried to store
into location8, which naturally failed.

This commit adds the /Index (starting object number) to the /Size before
ensuring the objects are large enough.

No differences expected