pdfwrite - handle CIDFonts with mixed type 3/type 1 descendants
Bug #691096 "the file generated by ps2pdf from this PS file does not display some chars"

The input has a CIDFont declared which uses both a type 1 and a type 3 font
as descendants. The pdfwrite device family can't handle this, resulting in
/.notdef glyphs.

In this commit; when we detect a switch to a type 3 descendant, and the previous
font was not a type 3, we set 'type3charpath' to true. This forces the
text handling to fall back to a bitmap font mode for this text.

This is less than ideal, but its a very rare occurence and does result in
all the text being rendered.

No differences expected