Fix to turn of BPC when we are using the Gray_to_K ICC profile.
This appears to be a bug in lcms.  Will follow up with Marti.   Also fix for tiff64nc.  In the test
tiff_compression_allowed we were missing the test for the 16 bit case.   Surprisingly this commit
results in progressions in the 8 bit device outputs too when we have a gray source and going
to a CMYK device.