pdfwrite - null a pointer to avoid an enumeration problem
pdev->vgstack wasn't being set to NULL when freed, which could lead to
garbage collection problems.

At the same time, remove two macros which I really don't like (and made this
problem difficult to locate) and set a couple more pointers to NULL when
the contents are freed.

Bug #694380, but unfortunately this simply causes a seg fault in a different
location.