pdfwrite - fix multi-plane images with < 8bpc
The code to pack multiple bit planes into bytes didn't properly account
for the number of bytes to be written to the output. We always wrote a
full block of bytes as long as the number of bytes remaining was not
greater than our block size. That is, it did not account for the fact that
we might not need to write a complete block of bytes, in the case where the
number of bytes left was exactly equal to the size of our internal block.

Fixed that by changing the '>=' to '>' so that if the size was equal we would
properly check to see how many bytes we actually need to write. This then
uncovered another problem, if the width of the image in bits was a multiple
of the block size, then we wouldn't write any data. Fixed that too.

No differences expected.