Make C 'param' collections free memory as per 'ref' params
Bug #693561

The 'memory corruption' referred to in the bug report is actually a memory
leak, and is only exposed by using the -Z? (check pointers) debugging switch.
As such its not as disastrous as the bug report makes it sound (I'm unable
to reproduce the SEGV).

Its caused because when we retrieve a parameter list from the device, we do
so using the 'C' param list structures rather than the PostScript 'ref'
params. One of the things that pdfwrite returns is image dictionaries, and
these require us to create a temporary param list. With the 'ref' param
structures, when we call param_end_write_dict (redirected through a macro
to param_end_write_collection, further redirected through a macro to
plist->procs->end_xmit_collection) the memory used by the list is explicitly
freed, when we use the 'C' param, it is *NOT* freed.

All the places in the code that I can find where we call param_end_write_dict
seem to assume that this will also free the memory. This commit modifies
the C param list code so that it also frees the memory. At least now the
two param types behave the same.

No differences expected