Repair performance degradation with PostScript functions
Bug #693561 "gs_c_param_list_release() corrupts memory"

The commit 289aa28f46e2be1c87c34d44ee86dec41e87506b removed the explicit
test of the device name with a test against a specifically named device
parameter. However, as noted in the bug, this led to a performance penalty
as the parameter was re-evaluated inside a loop.

This commit moves the parameter evaluation outside check_psc_function()
altogether as this routine can be called recursively. Instead the parameter
is retrieved in the calling routines and passed as a new argument.

No differences expected