ps2write - better documentation for HaveTrueTypes
The documentation for HaveTrueTypes stated that TrueType fonts would be
emitted as bitmap fonts if -dHaveTrueTypes is false. This information was
only partially correct, if the size of the bitmap (as determined by the
text size and resolution) exceeds the maximum for the cache then we don't
put the bitmap in a font at all, we emit it as an inline image.