txtwrite - fix a buffer overrun
Bug #695109 "gswin32c produces an access violation when using txtwrite"

A cut and paste error led to us trying to copy a char string as a sequence
of shorts, resulting in overrunning the source buffer. Depending on the
precise memory layout this could cause access to unallocated memory and
an application crash.

Thanks to José Antonio for persisting with this bug and doing all the
debugging work.

No differences expected.