pdfwrite - don't emit ICC spaces with no Alternate and no potential default alternate
Bug #691500 "Bad pdf file generated with pdfwrite"

The XPS file cotnains a 6 colour ICC profile, the XPS interpreter passes this
along as an ICCBased space, but does not construct an /Alternate space for
it. pdfwrite would then emit a PDF file with a 6 colour profile and no
/Alternate, which is illegal as there is no possible way to convert the
input into a device space without handling the ICC profile.

If we encounter this (which I think is only possible from XPS), we now give
up on the ICC profile and instead convert to the ProcessColorModel colour
space instead.

Ghostscript can read the resulting file without error, Acrobat gives a
'drawing error' and evince complains about input profiles, but in general
I'm going to label this as a fix.

No differences expected