pdfwrite - clamp CropBox to MediaBox for PDF/X output
Bug #695120 "Page boxes not nested properly when creating PDF/X using files with bleed"

The description of the bug is not accurate, the bleed is irrelevant.

The problem is caused because the specimen file has ridiculously accurate
MediaBox etc, specified to 3 decimal places. Due to rounding differences the
CropBox can be a tiny amount (thousandths of a millimeter) large than the
MediaBox.

I cannot find anywhere in the PDF/X-3 spec which says that the CropBox has
to be smaller than the MediaBox, but the Acrobat pre-flight tool complains
if it is.

This commit simply clamps the CropBox to the MediaBox that actually gets written
to the PDF file (for PDF/X only) and prevents the problem.

No differences expected