ps2write - fix CompressPages=false and SubFileDecode filter
commit c30fb930e7c38dd7bcf47b0305e77cb6ba003bc3 altered the default of the
CompressPages switch to false, this uncovered an interesting bug. The code
previously used the string 'endstream' as the EODString argument for a
SubFileDecode filter. When the stream contents are compressed this is reliable
but if the stream contents are not compressed they may contain the string
'endstream' (eg a document describing PDF syntax) which would cause premature
end of data detection.

In this commit we instead set the EODString to the empty string () and set
the EODCount to the length of the stream. This is more reliable, because we
trust the stream length we create ourselves, but does mean that the stream
can no longer be edited simply, the length must now be consistent with the
actual stream contents.

This fixes a bug introdcued with commit c30fb930e7c38dd7bcf47b0305e77cb6ba003bc3
involving the test file DML_003-2010-1_8.pdf