Bug 695087: modify jbig2 device to output jb2 files
The jbig2 device originally was written to output "bare" streams, as are written
to PDF files, but is of very limited use.

This commit changes it so the "bare" stream is written to a jb2 container file,
which can be easily opened, for example, by the jbig2dec command line tool.

Note that the output seems to be inverted, I haven't "fixed" that as I'm not
sure whether our jbi2 output is wrong, or the jbig2dec conversion to PNG is
wrong.

This is primarily to ease regression testing of the Luratech jbig2 encoder
integration.

No cluster differences.