pdfwrite - while processing pdfmarks, ensure strings are null-terminated before using sscanf
Bug #695119 "Ghostscript inconsistently fails creating links with pdfmark"

The problem turns out to be that while handling SrcPage in a pdfmark the
code executed sscanf on a string which was not guaranteed to be null-terminated

Here we copy the data and explicitly null terminate the copy before running
sscanf, so that the returned value is reliable.

Also remove an uninitialised variable that could potentially have caused a problem.

No differences expected.