PS interpreter - don't limit sampled tint transforms to 64k data
Bug #693237 "PDF restrictions on PS function"

The bug title is misleading, as is the comment in the report. The problem
has little to do with PDF restrictions. When we get a tint transform for a
complex space we convert it into a PostScript function rather than maintaining
the tint transform as PostScript (we do this for performance reasons).

In general we try to use a type 4 (PostScript calculator) function, as these
are the most flexible variety, but these are not defined for PostScript, only
for PDF, and it is not always possible to handle a tint transform as a type
4 function. In this case we fall back to using a type 0 (sampled) function.

In order to create a sampled function we must (obviously!) sample the original
function, and create a lookup table to store the samples. This is an M-in
N-out table, so as the number of inputs and outputs increases the size of
the table increases dramatically. Because we store the data in a string
the code checks the size of the lookup table against 64k and if it is too
large (as in this case) rejects it and tries again, taking fewer samples.

The minimum number of samples is 2 (we must sample each end of the ramp),
and if we cannot fit that into a 64k string, then we returned a rangecheck
error.

However we now permit strings greater than 64k, so while we still want to
keep the lookup table to a sensible size, we should not abort just because
the table is > 64k.

This commit alters the logic so that we try to fit into a 64k string, but
if we've reached the minimum number of samples we don't bother checking any
more and simply use a string big enough to hold the data.

This seems to work, however the example file is reduced to the point where
nothing is visible on the output, and although the bug report references
a customer file its not attached to the report, nor is there any other
information which would allow me to find the original file. So this isn't
hugely well tested in this unusual condition.

No differences expected.