Bug 695133: fix invalid types in form xobject matrix
The PDF in question has a name object rather a real or integer object as the
last entry in the matrix.

Before rendering the form, check the entries in the matrix, and if there is
an invalid type, replace it with a real value. I've opted to use 0.00 as the
replacement value, and this *seems* to reflect Acrobat's behavior. Creating
a test case where one of the scale values in the matrix is an invalid type
results in Acrobat not displaying the form, which hints at a degenerate matrix.
We have also previously established that Acrobat substitutes zero in other
similar cases.

No cluster differences.