Bug 695100: pxl jpeg decoding sometimes misses scans
jpeg decoding (s_DCTD_process) is used by all pdls (pdf/ps/pcl/pxl/xps)
but in different ways.
The PDLs provide an output buffer (pw->ptr) for the decoded scan(s), and
an input buffer (pr->ptr) with inputdata. XPS's outputbuffer can contain
the complete image, the other pdls provide a buffer for 1 scan.

So XPS calls the jpeg-decoder once for the complete image (including the
EOI marker), the other pdls call repeatedly for 1 scan, until all scans
have been retrieved.

The jpeg-decoder (s_DCTD_process) calls gs_jpeg_read_scanlines(inputdata)
to decode 1 scan (using the inputdata). If inputdata is not enough the pdl
is asked for more, until (atleast) 1 scan is available. This scan
(samples) is then copied into the outputbuffer (goto dumpbuffer:).
If the outputbuffer is not filled completely, the jpeg-decoder is asked
for another scan (i.e. for XPS) which is then also added to the
outputbuffer.
Else (if the outputbuffer cannot contain the next scan) it is passed back
to the graphics-lib (pl_image_data), the outputbuffer is emptied and the
jpeg-decoder is called for the next scan.

However, with PXL, provided outputbuffer is exactly the size of a scan.
This means that it is only transferred, when s_DCTD_process wants to copy
a next scan (which fails because the outputbuffer is full; the
transferring is delayed), or end-of-data is reached.
If however, the gs_jpeg_read_scanlines() needs multiple inputbuffers to
provide this next scan, then this next scan will be copied over the not
yet transferred scan, thus causing the previous scan to be missed.

The fix in sdctd.c prevents the delaying and copies immediately.
The mod in pximage.c ensures that the complete imagedata (incl the
jpeg-EOI) is consumed (otherwise a bufferoverflow is triggered).

Some (PXL) cluster differences (progressions).