PXL font code - correct the order of a matrix multiplication
Bug #693947 "The Intellifont as type 3 font commit causes a regression."

Matrix multiplication is not commutative, and the order of operands was
incorrect in this case. Since it only matters for high level devices it
naturally only showed up with pdfwrite.

This shows a progression (with pdfwrite) in C207.bin and bug691464b.xl
which actually appears to be the same file.