Bug 692577: tweak lprsetup.sh for non-default locations
Provide a command line option to lprsetup.sh to specify where the filter files
should be written. Also, implement a simple heuristic to attempt to identify
the correct location for the filter files.

Finally, if we rely on the above heuristic, issue a warning and give the user
the chance to abort the script.

CLUSTER_UNTESTED