Bug 695138: PCL pushcursor in definemacro not stored correctly (combined command)
When a macro is defined with a combined command and de define is not
the closing command (i.e. 0x) then the continuation as stored in the
macro_prefix (pcparse.c) is not correct. It contains only the <ESC>&,
but should have been <ESC>&f).
Thus a following 0S is stored as <ESC>&0S, i.s.o. <ESC>&f0S, and is not
recognized as a valid push-cursor command.