Disable the remaining long form color commands, fixing 695110.
All of the color long form commands appear to be no longer supported
on modern HP devices and our implementation of them frequently results
in unexpected output and difficult to find bugs.  This change gets rid
of the remaining commands which affected white and black point
settings.  Aside from fixing the test file in the bug this improves
several panels in the Quality Logic FTS against the Color Laserjet
4700, i.e. panels 810 and 814.

The command is simply truncated so the long form command is
transformed into a corresponding short form allowing for setting of
the color space, bits per component, etc.  The long form code remains
for now and is easily brought back if we get user feedback that there
is an HP device we care to support with these extended color commands.
If that is the case it will have to be handled with conditional
compilation or a device emulation mode.  We'd like to avoid that if
possible.