Fix 695417 cursor tracking broken.
The bug subject is incorrect, the problem was initially diagnosed as a
problem with GL/2 and PCL cursor interaction, but in fact it was the
rounding of the translation components of PCL raster to device space
matrix produced incorrect results at some resolutions.

This problem needs a bit more investigation but empirically not
rounding translation components seems to give better results.  I don't
know if this is because it more closely emulates HP's calculations or
if there is another interaction in the ghostscript graphics library
which makes the integral translation components inappropriate.

The change results in many single pixel translations in the regression
tests.