PDF interpreter - sunthesise appearances for FreeText annotations
Bug #693731 "Freetext annotations are not printed"

The FreeText annotation in question has no appearance stream and we did
no previously synthesise one. This commit adds code to do so, it should
work with UTF16-BE text but I don't have a test file for this.

The existing code to draw annotation borders was insufficient, it did not
correctly drop elements drawn with a colour of /C[] and would not handle
colours (/C) defined in anything except RGB space.

The Sumatra test file annotations_galore_II.pdf exhibits a progression
with this commit (its the only other example of this I could find)