PDF interpreter - improve transparency detection and apply it to annotations
The existence of CA and ca in annotations without appearance streams were
not detected. The FreeText annotation synthesis can apply transparency if
present, so we need to detect it and push the compositor.

Also, add the code to the FreeText annotation appearance synthesis to apply
transparency.

A few (3) Sumatra PDF files now render very slightly differently, and the
same files now produce larger PostScript than they did before (because they
are rendered to images)