PDF interpreter - reinstate (and complete) the BMC/BDC/EMC pdfmark support
Bug #695151 "Add BMC/EMC and BDC pdfmark support"

This is the patch supplied by Alexander Grahn with a couple of very minor
additions. Firstly we clean up the Properties resource in pdf_close in case
by some chance there are still properties dictionaries lying around. Secondly
we treat the Properties resource specially (like fonts) in
pdf_store_page_resources().

Because we write the properties dictionaries in line, we must not emit them
in pdf_store_page_resources (they no longer exist for one thing) as this
leads to duplicate objects being created. We may need to revisit this in the
future as I'm not 100% certain this is the best way to proceed. However it
works properly for all the (very few) test cases I have, or have been able
to create.

This patch also 'fixes' the opdfread ps2write header. In the past if a BMC
EMC or BDC was encountered in the PDF file it would try to handle it, but
the handler doesn't exist, leading to PostScript errors. Rather than try
to reconstruct the requisite pdfmarks I've chosen simply to ignore these
constructs. If anyone objects strenuously we can figure out a handler
in future.

No differences expected.