Update Ghostscript solution for Windows.
Later versions of VS dislike relative paths for the working directory
in the debug set up. The solution is to use "$(ProjectDir)\.." as this
evaluates to the same dir as ".." used to. Update the ReadMe to
reflect this.

Also update the Solution files so that "Output" points to the built
executables. This enables the editing required for the debugging
setup to be simpler.