pdfwrite - fix Named destinations with Names tree
Bug #695162 "Internal hyperlinks lost when converting to pdf"
Bug #695175 "can't save PDF file if compiled with hyperref packages"

In order to add support for the EMBED pdfmark we had to move the named destinations
from the PDF 1.1 /Dests dictionary to the PDF 1.2+ Names tree. The Names
tree requires that the 'keys' be strings, unlike the old Dests dict where
the keys are of course name objects.

Now the PDF reference says that,with PDF 1.2 and above, both strings and
name objects can be used to refer to named destinations. However it appears
that this is not true for Acrobat, the reference to a named destination must
be a string object if we are using a Names tree.

This commit alters the creation of the Names tree so that we store the actual
destinations as individual objects (this probably wasn't required, but it
mimics Distiller) instead of inline dictionaries, reinstates a missing /D
key in said dictionary and crucially alters the reference to the named
destination so that it is a string and not a name.

The commit also caters for 'unusual' names in named destinations. The PDF
interpreter 'escapes' the name of a name object by adding leading NULLs if
the name contains certain kinds of characters. We weren't catering for that
when converting names to strings for inclusion in the Names tree, which was
leading to an empty object, and caused Acrobat to refuse to save the file.

Finally, while investigating the problems, I noticed a logical error with
DOCVIEW pdfmarks which could result in Open actions being written incorrectly
and this is also fixed here.

No differences expected as the cluster doesn't test this kind of metadata.