Add support for changing the default device list at runtime
(Bug 695096)

The graphics library searches a list of preferred devices to find a suitable
default device. Previously that list was created at graphics library compile
time and could not be changed after that.

This change allows the calling application to retrieve the current string that
represents this list, and set its preferred list and order, if required.

No cluster differences.