Implement use of the new 'device param' spec op
Wherever possible, replace the .getdeviceparams call with a call to the
PostScript .special_op to retrieve a single parameter.

This cannot be done unless we can guarantee that the devices we are
interested in will implement the spec_op device method. So some places
which appear at first sight to be candidates can't be done this way.