Fix some scan-build warnings from commit 0476b5036d8175a4817a9385cb6a15b64fab310e
The code dealing with single device parameters was over-aggressive about
reducing the ICC profile setup. In practice this is unlikely ever to be
a problem, there is no real chance that the setup won't be performed
before we reach this point, but its obviously much better to have the code
correct.

This commit essentially copies verbatim the setup code from the normal
get_params device method in the default device.